+++ include/vulkan/vk_layer.h
@@ -234,29 +234,6 @@ typedef struct VkLayerInstanceDispatchTable_ {
         CreateDisplayPlaneSurfaceKHR;
 } VkLayerInstanceDispatchTable;
 
-
-// LL node for tree of dbg callback functions
-typedef struct VkLayerDbgFunctionNode_ {
-    VkDebugReportCallbackEXT msgCallback;
-    PFN_vkDebugReportCallbackEXT pfnMsgCallback;
-    VkFlags msgFlags;
-    void *pUserData;
-    bool default_callback;
-    struct VkLayerDbgFunctionNode_ *pNext;
-} VkLayerDbgFunctionNode;
-
-// TODO:  Move debug action definitions out of this header file
-// Definitions for Debug Actions
-typedef enum VkLayerDbgActionBits {
-    VK_DBG_LAYER_ACTION_IGNORE = 0x00000000,
-    VK_DBG_LAYER_ACTION_CALLBACK = 0x00000001,
-    VK_DBG_LAYER_ACTION_LOG_MSG = 0x00000002,
-    VK_DBG_LAYER_ACTION_BREAK = 0x00000004,
-    VK_DBG_LAYER_ACTION_DEBUG_OUTPUT = 0x00000008,
-    VK_DBG_LAYER_ACTION_DEFAULT = 0x40000000,
-} VkLayerDbgActionBits;
-typedef VkFlags VkLayerDbgActionFlags;
-
 // ------------------------------------------------------------------------------------------------
 // CreateInstance and CreateDevice support structures
